\name{DownUpMatrix}
\alias{.required}
\alias{corr.from.pi}
\alias{DownUpMatrix}
\alias{Marginals}
\alias{mc.est.raw}
\alias{mChoose}
\alias{mChooseTable}
\alias{p.from.tau}
\alias{tau.from.pi}
\title{Internal CorrBin objects}
\usage{
  DownUpMatrix(size, ntrt, turn)

  mChoose(n, rvec, log = FALSE)

  Marginals(theta)

  mc.est.raw(object, ...)

  tau.from.pi(pimat)

  p.from.tau(taumat)

  corr.from.pi(pimat)

  mChooseTable(n, k, log = FALSE)
}
\description{
  Internal CorrBin objects.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}

