% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBData.R
\name{read.CBData}
\alias{read.CBData}
\title{Read data from external file into a CBData object}
\usage{
read.CBData(file, with.freq = TRUE, ...)
}
\arguments{
\item{file}{name of file with data. The first column should contain the
treatment group, the second the size of the cluster, the third the number of
responses in the cluster. Optionally, a fourth column could give the number
of times the given combination occurs in the data.}

\item{with.freq}{logical indicator of whether a frequency variable is present
in the file}

\item{...}{additonal arguments passed to \code{\link[utils]{read.table}}}
}
\value{
a \code{CBData} object
}
\description{
A convenience function to read data from specially structured file directly
into a \code{CBData} object.
}
\seealso{
\code{\link{CBData}}
}
\author{
Aniko Szabo
}
\keyword{IO}
\keyword{file}
