% Generated by roxygen2 (4.0.2): do not edit by hand
\name{[.CMData}
\alias{[.CMData}
\title{Extract from a CMData object}
\usage{
\method{[}{CMData}(x, i, j, drop)
}
\arguments{
\item{x}{\code{CMData} object.}

\item{i}{numeric, row index of extracted values}

\item{j}{numeric, column index of extracted values}

\item{drop}{logical. If TRUE the result is coerced to the lowest possible dimension.
The default is the same as for \code{\link{[.data.frame}}: to drop if only one column is left, but not to drop if only one row is left.}
}
\value{
a \code{CMData} object
}
\description{
The extracting syntax works as for \code{\link{[.data.frame}}, and in general the returned object is not a \code{CMData} object.
However if the columns are not modified, then the result is still a \code{CMData} object  with appropriate attributes  preserved,
and the unused levels of treatment groups dropped.
}
\author{
Aniko Szabo
}
\seealso{
\code{\link{CMData}}
}
\keyword{IO}
\keyword{file}

