% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchMultinomial.R, R/Reprod.R
\name{mChoose}
\alias{mChoose}
\alias{Marginals}
\alias{mc.est.raw}
\alias{tau.from.pi}
\alias{p.from.tau}
\alias{corr.from.tau}
\alias{corr.from.pi}
\alias{mChooseTable}
\alias{DownUpMatrix}
\alias{.required}
\title{Internal CorrBin objects}
\usage{
mChoose(n, rvec, log = FALSE)

Marginals(theta)

mc.est.raw(object, ...)

tau.from.pi(pimat)

p.from.tau(taumat)

corr.from.tau(taumat)

corr.from.pi(pimat)

mChooseTable(n, k, log = FALSE)

DownUpMatrix(size, ntrt, turn)
}
\description{
Internal CorrBin objects.
}
\details{
These are not to be called by the user.
}
\keyword{internal}
